+++ src/main/java/pl/betoncraft/betonquest/compatibility/worldguard/RegionCondition.java
@@ -1,13 +1,5 @@
 package pl.betoncraft.betonquest.compatibility.worldguard;
 
-
-import com.sk89q.worldedit.bukkit.BukkitAdapter;
-import com.sk89q.worldedit.math.BlockVector3;
-import com.sk89q.worldguard.WorldGuard;
-import com.sk89q.worldguard.internal.platform.WorldGuardPlatform;
-import com.sk89q.worldguard.protection.ApplicableRegionSet;
-import com.sk89q.worldguard.protection.managers.RegionManager;
-import com.sk89q.worldguard.protection.regions.ProtectedRegion;
 import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.Instruction;
 import pl.betoncraft.betonquest.api.Condition;
@@ -30,19 +22,7 @@ public class RegionCondition extends Condition {
     @Override
     protected Boolean execute(final String playerID) {
         final Player player = PlayerConverter.getPlayer(playerID);
-        final WorldGuardPlatform worldguardPlatform = WorldGuard.getInstance().getPlatform();
-        final RegionManager manager = worldguardPlatform.getRegionContainer().get(BukkitAdapter.adapt(player.getWorld()));
-        if (manager == null) {
-            return false;
-        }
-        final ProtectedRegion region = manager.getRegion(name);
-        final ApplicableRegionSet set = manager.getApplicableRegions(BlockVector3.at(player.getLocation().getX(), player.getLocation().getY(), player.getLocation().getZ()));
-        for (final ProtectedRegion compare : set) {
-            if (compare.equals(region)) {
-                return true;
-            }
-        }
-        return false;
+        return WorldGuardIntegrator.isInsideRegion(player.getLocation(), name);
     }
 
 }